@@ -1,7 +1,7 @@
 package com.qiangxi.demo.viewpostdemo;
 
 import android.content.Context;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.util.AttributeSet;
 import android.widget.TextView;
 
