@@ -1,6 +1,6 @@
 package com.qiangxi.demo.viewpostdemo;
 
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
 
